+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  * Moved most settings to sync storage
  * Changed options page to use port messaging, so having the options page doesn't cause the addon to stop working
  * Made the progress bar for importing dictionaries actually show the progress
+ * Added a volume option for playing sound
 
 #### Version 1.0.18
  * Fixed broken Anki Import